refactor: media query 사이즈 theme 에 등록하여 사용하도록 수정
